-updates to Kafka 3.2 -updates to SpringBoot 2.7.0 -updates to Java 18 -simplifies code to focus on consuming Kafka only -moves build to maven
