+++ .gitignore
@@ -1,49 +1,22 @@
-# Created by .ignore support plugin (hsz.mobi)
-
-### JetBrains template
-# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion
-
 *.iml
-
-## Directory-based project format:
 .idea/
-# if you remove the above rule, at least ignore the following:
-
-## File-based project format:
 *.ipr
 *.iws
-
-## Plugin-specific files:
-
-# IntelliJ
 /out/
-
-# mpeltonen/sbt-idea plugin
+/target
 .idea_modules/
-
-# JIRA plugin
-atlassian-ide-plugin.xml
-
-# Crashlytics plugin (for Android Studio and IntelliJ)
-com_crashlytics_export_strings.xml
-crashlytics.properties
-crashlytics-build.properties
-
-
-### Gradle template
 .gradle
 build/
-
-# Ignore Gradle GUI config
 gradle-app.setting
-
-# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
 !gradle-wrapper.jar
-
-# Package Files #
 *.jar
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
-
-
+*.class
+*.log
+.mtj.tmp/
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
