--- build.gradle
@@ -1,48 +1,0 @@
-group 'de.mobile'
-version '1.0-SNAPSHOT'
-
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:1.2.3.RELEASE")
-    }
-}
-
-apply plugin: 'java'
-apply plugin: 'spring-boot'
-
-compileJava.dependsOn(processResources)
-
-jar {
-    baseName = 'kafka-picture-consumer'
-    version = '0.1.0'
-}
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-repositories {
-    repositories {
-        mavenCentral()
-    }
-}
-
-dependencies {
-    compile(group: 'org.springframework.boot', name: 'spring-boot-starter-parent')
-    compile(group: 'org.springframework.boot', name: 'spring-boot-starter')
-    compile(group: 'org.apache.kafka', name: 'kafka_2.10', version: '0.8.2.1') {
-        exclude module: 'slf4j-log4j12'
-    }
-    compile(group: 'com.google.code.gson', name: 'gson', version: '2.3.1')
-    compile(group: 'com.google.guava', name: 'guava', version: '18.0')
-    compile(group: 'javax.inject', name: 'javax.inject', version: '1')
-    compile(group: 'io.dropwizard.metrics', name: 'metrics-core', version: '3.1.2')
-
-    testCompile group: 'junit', name: 'junit', version: '4.11'
-}
-
-wrapper {
-    gradleVersion = '2.4'
-}
