--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu May 21 21:07:01 CEST 2015
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.2-all.zip
