+++ src/main/java/de/opitz/sample/kafka/consumer/Application.java
@@ -1,16 +1,13 @@
-package run;
+package de.opitz.sample.kafka.consumer;
 
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EnableConfigurationProperties
-@ComponentScan (basePackages = {"config", "run", "kafka", "ui"})
 public class Application {
-
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
