+++ src/main/java/de/opitz/sample/kafka/consumer/ui/Image.java
@@ -1,0 +1,6 @@
+package de.opitz.sample.kafka.consumer.ui;
+
+import java.awt.image.BufferedImage;
+
+record Image(String name, BufferedImage bufferedImage) {
+}
