--- src/main/resources/application.properties
@@ -1,3 +1,0 @@
-zookeeper.connect=LM-TXL-00677741:2181
-kafka.group.id=images
-kafka.topic=images
