+++ startConsumerGroup3.sh
@@ -1,1 +1,1 @@
-java -Djava.awt.headless=false -jar build/libs/kafka-picture-consumer-0.1.0.jar --kafka.host=localhost:9092 --kafka.group.id=3 &
+java -Djava.awt.headless=false -jar target/kafka-picture-consumer-2.0.0.jar --ui.position.top=700 --kafka.group.id=3 &
